requirements.txt
astroid==2.2.5
automium==0.2.6
automium-web==0.1.1
Click==7.0
colorama==0.4.1
coverage==4.5.4
Django==2.2.5
django-chartjs==1.5.0
django-crispy-forms==1.7.2
django-email-users==1.1.0
django-nose==1.4.6
django-tables2==2.2.1
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
gunicorn==19.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
legacy==0.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
neat-python==0.92
nose==1.3.7
paho-mqtt==1.5.0
Pillow==6.1.0
pony==0.7.11
pycodestyle==2.5.0
pyflakes==2.1.1
pygame==1.9.6
pylint==2.3.1
pytz==2019.2
rednose==1.3.0
six==1.12.0
sqlparse==0.3.0
store==2020.1.3
termstyle==0.1.11
typed-ast==1.4.0
users==1.0.dev0
views==0.3
Werkzeug==0.15.6
wrapt==1.11.2
